"Angular Application Bootstrap "
This code snippet serves as the launchpad for the Angular application. It employs the platformBrowserDynamic module to dynamically bootstrap the AppModule, which acts as the central module for the application.

By invoking the bootstrapModule function, the Angular framework initiates the loading and rendering of components and modules, kickstarting the application's lifecycle. Any initialization errors are captured and displayed in the console.

This concise piece of code ensures the smooth and efficient initiation of the "Todo" Angular application, laying the foundation for a responsive and interactive user experience.